refactor: rename files to snake_case
