+++ src/application/auth/services/sign_in_user/helpers/create_jwt_token.go
