+++ src/application/auth/services/sign_in_user/helpers/find_by_email.go
